Artem fixed model

+++ FinanceManager/DAL/Models/Currency.cs
@@ -5,8 +5,8 @@ namespace FinanceManager.DAL.Models
     public class Currency : BaseModel
     {
         [PrimaryKey]
-        public int CurrencyId { get; set; }
+        public short CurrencyId { get; set; }
         public string Name { get; set; }
-        public int Value { get; set; }
+        public double Value { get; set; }
     }
 }
